temp_todelete/requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
python-dotenv>=0.19.0
pymysql>=1.0.0
cryptography>=41.0.0  # Required for PyMySQL with SSL
PyJWT>=2.0.0  # Added for authentication
ijson>=3.2.0  # Added for iterative JSON parsing
